refactor: refactor field and add more opts
